\name{ExampleRNAseq}
\docType{data}
\alias{ExampleRNAseq}
\title{RNAseq example dataset}
\description{
  A simulated example dataset of RNAseq for PQLseq.
}
\usage{data(ExampleRNAseq)}
\format{
  Contains the following objects:
  \describe{
  	\item{count}{a data frame containing the read counts for 5 genes.}
    \item{predictor}{a vector of 100 observations on a continuous variable.}
    \item{relatednessmatrix}{a genetic relationship matrix for 100 individuals.}
    \item{totalcount}{a data frame containing the total read counts for 5 genes.}
  }
}
\keyword{datasets}
